v0.7.0-pre.3 - extract LitVirtualizer for side-effect-free imports
